fix(web-stories-wp): stackable inputs component composability and styling enhancements
Closes #webstories-9461
