+++ packages/story-editor/src/components/form/index.js
@@ -31,4 +31,3 @@ export {
   default as HierarchicalInput,
   makeFlatOptionTree,
 } from './hierarchical';
-export { default as StackableGroup } from './stackable';
