+++ packages/story-editor/src/components/form/stackable/index.js
@@ -13,52 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * External dependencies
- */
-import PropTypes from 'prop-types';
-import styled, { css } from 'styled-components';
 
-/**
- * Internal dependencies
- */
-
-const StackableContainer = styled.div(
-  ({ locked, styleOverride }) => css`
-    display: flex;
-    ${locked ? 'max-width: 106px' : ''}
-    ${styleOverride};
-
-    & > div {
-      border-radius: 0;
-
-      &:first-of-type > div {
-        border-radius: 4px 0 0 4px;
-      }
-      &:last-of-type > div {
-        border-radius: 0 4px 4px 0;
-      }
-    }
-  `
-);
-
-function StackableGroup(props) {
-  const { children, containerStyleOverride, locked } = props;
-
-  return (
-    <StackableContainer locked={locked} styleOverride={containerStyleOverride}>
-      {children}
-    </StackableContainer>
-  );
-}
-
-StackableGroup.propTypes = {
-  children: PropTypes.node,
-  locked: PropTypes.bool,
-  containerStyleOverride: PropTypes.oneOfType([
-    PropTypes.object,
-    PropTypes.array,
-  ]),
-};
-
-export default StackableGroup;
+export { default as StackableGroup } from './stackableGroup';
+export { default as StackableInput } from './stackableInput';
